implement chat, message commands, refactor chat command naming
+ avatar user command
+ translation message command, replacing reaction translations

+++ src/main/kotlin/moe/kabii/data/mongodb/guilds/MusicSettings.kt
@@ -1,7 +1,7 @@
 package moe.kabii.data.mongodb.guilds
 
 data class MusicSettings(
-    var startingVolume: Int = defaultStartingVolume,
+    var startingVolume: Long = defaultStartingVolume.toLong(),
     var lastChannel: Long? = null,
     var deleteOldBotMessages: Boolean = true,
     var queuerFSkip: Boolean = true,
